Refactored: Made DateTime conversions into a concern. RequestHelper is namespaced within Booker and is included within classes, enabling it to be more easily shared amongst clients
