+++ lib/booker/v4.1/customer.rb
@@ -1,7 +1,7 @@
 module Booker
   module V41
     class Customer < Booker::Client
-      include Booker::V4::RequestHelper
+      include ::Booker::RequestHelper
 
       V41_PREFIX = '/v4.1/customer'
       V41_APPOINTMENTS_PREFIX = "#{V41_PREFIX}/appointment"
