+++ lib/booker/v4/common_rest.rb
@@ -1,7 +1,7 @@
 module Booker
   module V4
     module CommonREST
-      include Booker::V4::RequestHelper
+      include ::Booker::RequestHelper
 
       def get_online_booking_settings(booker_location_id:)
         response = get("/location/#{booker_location_id}/online_booking_settings", build_params)
