+++ lib/booker/v5/availability.rb
@@ -1,6 +1,8 @@
 module Booker
   module V5
     class Availability < Booker::Client
+      include ::Booker::RequestHelper
+
       API_METHODS = {
         availability: '/v5/availability/availability'.freeze,
         two_day_availability: '/v5/availability/2day'.freeze,
