+++ lib/booker/v5/request_helper.rb
@@ -1,0 +1,6 @@
+module Booker
+  module V5
+    include ::Booker::RequestHelper
+
+  end
+end
