+++ lib/booker_ruby.rb
@@ -16,6 +16,10 @@ require 'booker/booker'
 # Errors
 require 'booker/errors'
 
+# Concerns
+require 'booker/concerns/date_time_concern'
+require 'booker/request_helper'
+
 # Models
 require 'booker/model'
 
@@ -23,6 +27,7 @@ require 'booker/model'
 require 'booker/v4/models/model'
 
 # Types
+require 'booker/v4/request_helper'
 require 'booker/v4/models/type'
 require 'booker/v4/models/country'
 require 'booker/v4/models/status'
@@ -83,6 +88,7 @@ require 'booker/v4/models/notification_settings'
 require 'booker/v4/models/feature_settings'
 
 # V5 Models
+require 'booker/v5/request_helper'
 require 'booker/v5/models/model'
 require 'booker/v5/models/location_hour'
 require 'booker/v5/models/availability'
@@ -95,7 +101,6 @@ require 'booker/v5/models/availability_result'
 require 'booker/client'
 
 # V4 Rest
-require 'booker/v4/request_helper'
 require 'booker/v4/common_rest'
 require 'booker/v4/business_rest'
 require 'booker/v4/customer_rest'
@@ -106,6 +111,7 @@ require 'booker/generic_token_store'
 # Client Subclasses
 require 'booker/v4/business_client'
 require 'booker/v4/customer_client'
+# require 'booker/v4.1/request_helper'
 require 'booker/v4.1/customer'
 require 'booker/v4.1/merchant'
 require 'booker/v5/availability'
